interface and Default implementation for several classes
